feat(graphql): add new GraphQL operations for labels, PRs, and repository
- Add `create_label.rs` for creating labels via GraphQL
- Add `get_label_id.rs` for retrieving or creating label IDs
- Rename `prs.rs` to `get_open_prs.rs` and update trait to `GraphQLGetOpenPRs`
- Add `get_pr_id.rs` for retrieving pull request IDs
- Rename `pr_title.rs` to `get_pr_title.rs`
- Rename `repo.rs` to `get_repo_id.rs` and update trait to `GraphQLGetRepoID`
- Delete `label.rs` and split functionality into separate files
- Add `label_pr.rs` for adding labels to pull requests
- Update `mod.rs` to include new modules and traits
- Update `client/mod.rs` to include new headers for GitHub GraphQL client
- Update `git_ops.rs` to use new `GraphQL
