+++ src/lib/client/mod.rs
@@ -210,7 +210,11 @@ impl Client {
 
         let auth = format!("Bearer {}", token);
 
-        let headers = HashMap::from([("User-Agent", owner), ("Authorization", &auth)]);
+        let headers = HashMap::from([
+            ("User-Agent", owner),
+            ("Authorization", &auth),
+            ("X-Github-Next-Global-ID:", "1"),
+        ]);
 
         let github_graphql = gql_client::Client::new_with_headers(END_POINT, headers);
